requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
SQLAlchemy>=2.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-docx>=1.1.0
